app/requirements.txt
numpy==1.16.6
tensorflow==2.2.0
pillow==7.0.0
streamlit==0.61.0